requirements.txt
flask
gunicorn
psycopg2-binary
python-dotenv
feedparser
pyyaml
requests
beautifulsoup4
python-dateutil
